Add css module for pricematch component, refactored to remove price component
