+++ src/components/Fullfillment/style.css
@@ -1,12 +1,10 @@
 body {
   font-family: Arial,Helvetica,sans-serif;
   font-weight: 400;
-
-
-
 }
 /* Fullfillment.js component border  remove bottom padding later*/
 .pricingFullfillment {
+  margin-block-start: 24px;
   border-top-style: solid;
   padding-top: 24px;
   border-bottom-width: 0;
@@ -17,7 +15,7 @@ body {
   border-top-width: 1px;
   border-bottom-color: rgb(197, 203, 213);
   border-bottom-style: solid;
-  border-bottom-width: 1px;
+  border-bottom-width: 0px;
   /* align-content: space-between; */
   /* flex-direction: column; */
   /* box-sizing: border-box; */
@@ -25,7 +23,7 @@ body {
 }
 .pickup {
     color: rgb(48, 136, 0);
-    font-size: 12px;
+    font-size: 13px;
 
 }
 
@@ -35,7 +33,7 @@ body {
   color: #0457c8;
   text-decoration: none;
   text-decoration-line: none;
-  font-size: 12px;
+  font-size: 13px;
   cursor: pointer;
 }
 
@@ -50,13 +48,13 @@ body {
 }
 .availability {
   margin-top: 4px;
-  font-size: 12px;
+  font-size: 13px;
   font-weight: 400px
 }
 
 .impUpdateTitle {
   margin-top: 4px;
-  font-size: 12px;
+  font-size: 13px;
   font-weight: bold;
   color: black;
 }
@@ -67,7 +65,7 @@ body {
   color: #0457c8;
   text-decoration: none;
   text-decoration-line: none;
-  font-size: 12px;
+  font-size: 13px;
   cursor: pointer;
 }
 
@@ -85,8 +83,10 @@ body {
   color: #0457c8;
   text-decoration: none;
   text-decoration-line: none;
-  font-size: 12px;
+  font-size: 13px;
   cursor: pointer;
+  display: block;
+  width: 5%
 }
 
 .shippingZip:hover {
@@ -94,10 +94,3 @@ body {
   text-decoration: underline;
 }
 
-.title6 {
-  color: blueviolet;
-}
-
-.title7 {
-  color: blueviolet;
-}
