--- src/components/Price.js
@@ -1,12 +1,0 @@
-import React from 'react';
-
-const Price = (props) => {
-//use hooks here
-  return (
-    <div id="pricing-price">
-      {`$${props.pricing}`}
-    </div>
-  )
-
-}
-export default Price;
