+++ src/components/PriceMatch/style.css
@@ -1,0 +1,75 @@
+
+
+.priceMatchContainer {
+
+  padding-top: 24px;
+  border-bottom-width: 0;
+  border-left-width: 0;
+  border-right-width: 0;
+  border-top-color: rgb(197, 203, 213);
+
+  border-top-width: 1px;
+  border-bottom-color: rgb(197, 203, 213);
+  border-bottom-style: solid;
+  border-bottom-width: 0px;
+  /* align-content: space-between; */
+  /* flex-direction: column; */
+  /* box-sizing: border-box; */
+  padding-bottom: 0px;
+}
+
+.priceMatchBadge {
+  color: #0457C8;
+  background-image: url(https://assets.bbystatic.com/pricing/price/dist/client/images/c801f65e2b998a6c109b9aa2b5336a02.svg);
+  background-repeat: no-repeat;
+  background-size: contain;
+  display: inline-block;
+  padding-left: 14.9px !important;
+  font-size: 11px;
+  font-weight: 500;
+  border: none;
+}
+.priceMatchBadge:hover {
+  cursor: pointer;
+}
+.dollarSign {
+  font-stretch: ultra-condensed;
+  font-weight: 1000px;
+  font-size: 25px;
+  line-height: 35px;
+  font-style: normal;
+  align-items: center;
+}
+.pricing {
+  display: flex;
+  font-size: 30px;
+  line-height: 35px;
+  font-family: Arial, Helvetica, sans-serif;
+  font-weight: bold;
+  font-style: normal;
+  font-stretch: normal;
+  color: #1d252c;
+  box-sizing: border-box;
+}
+.priceMatcherContainer {
+  font-family: "Human BBY Web", Arial, Helvetica, sans-serif;
+  font-size: 13px;
+  line-height: 24px;
+  font-weight: 400;
+  font-style: normal;
+  font-stretch: normal;
+  display: inline-block;
+}
+
+.matchedPrice {
+  font-family: Arial, Helvetica, sans-serif;
+  font-weight: 700;
+  font-style: normal;
+  font-stretch: normal;
+  display: inline-block;
+  letter-spacing: -0.17px;
+  margin-right: 4px;
+  color: #fff;
+  background-color: #bb0628;
+
+}
